requirements.txt
streamlit
google-generativeai
python-docx
python-dotenv
PyPDF2